Improved database sync Apply default function when creating a column
